From and backend handling works (for now)

+++ Models/JSONmodels/routeInfo.cs
@@ -3,5 +3,5 @@ namespace CosmosOdyssey.Models;
 public class routeInfo {
     public from from { get; set; }
     public to to { get; set; }
-    public int distance { get; set; }
+    public Int64 distance { get; set; }
 }
